refactor(apex-replay-debugger): extract createStringFromVarContainer to shared module - W-11173323
